requirements.txt
fastapi
uvicorn[standard]
httpx
weaviate-client>=3.26.7,<4.0.0
pydantic
python-dotenv
